+++ app/src/main/java/com/nj/baijiayun/imageloader/test/MainActivity.java
@@ -1,23 +1,27 @@
 package com.nj.baijiayun.imageloader.test;
 
 import android.content.Intent;
-import android.graphics.drawable.BitmapDrawable;
+import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
-import com.nj.baijiayun.imageloader.listener.LoadListener;
+import com.bumptech.glide.request.transition.Transition;
 import com.nj.baijiayun.imageloader.loader.ImageLoader;
+import com.nj.baijiayun.imageloader.target.BitmapTarget;
 
 public class MainActivity extends AppCompatActivity {
 
     private static final String TAG = "MainActivity";
     public static String url = "https://upload.jianshu.io/users/upload_avatars/4951294/34667d56-03fa-4d00-bd68-81433d73f7de.png?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96";
-public static String url2="https://www.baidu.com/img/superlogo_c4d7df0a003d3db9b65e9ef0fe6da1ec.png?where=super";
+    public static String url2 = "https://www.baidu.com/img/superlogo_c4d7df0a003d3db9b65e9ef0fe6da1ec.png?where=super";
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -37,68 +41,97 @@ public static String url2="https://www.baidu.com/img/superlogo_c4d7df0a003d3db9b
 //
 
 
-        ImageLoader
-                .with(this)
-                .load(url2)
-                .rectRoundCorner(1)
-                .openBlur()
-                .asSquare()
-                .placeHolder(R.mipmap.common_course_dufault)
-                .listener(new LoadListener() {
-                    @Override
-                    public boolean onSuccess(Object resource) {
-                        Log.e("===", "onSuccess" + resource);
-                        if(resource instanceof BitmapDrawable)
-                        {
-                            ImageView viewById = findViewById(R.id.img2);
-                            viewById.setImageDrawable((Drawable) resource);
-                        }
+//        ImageLoader
+//                .with(this)
+//                .load(url2)
+//                .rectRoundCorner(1)
+//                .openBlur()
+//                .asSquare()
+//                .placeHolder(R.mipmap.common_course_dufault)
+//                .listener(new LoadListener() {
+//                    @Override
+//                    public boolean onSuccess(Object resource) {
+//                        Log.e("===", "onSuccess" + resource);
+//                        if (resource instanceof BitmapDrawable) {
+//                            ImageView viewById = findViewById(R.id.img2);
+//                            viewById.setImageDrawable((Drawable) resource);
+//                        }
+//
+//                        return true;
+//                    }
+//
+//                    @Override
+//                    public boolean onFail(Exception e) {
+//                        Log.e("===", "onFail" + e.getMessage());
+//
+//                        return false;
+//                    }
+//
+//                    @Override
+//                    public void preLoad() {
+//                        Log.e("===", "preLoad");
+//
+//                    }
+//                }).into((ImageView) findViewById(R.id.img3));
 
-                        return true;
-                    }
+//        ImageLoader.with(this)
+//                .load(url)
+//                .rectRoundCorner(1)
+//                .openBlur()
+//                .asCircle()
+//                .placeHolder(R.mipmap.common_course_dufault)
+//                .listener(new LoadListener() {
+//                    @Override
+//                    public boolean onSuccess(Object resource) {
+//
+//                        Log.e("===>>>", "onSuccess" + resource);
+//                        return false;
+//                    }
+//
+//                    @Override
+//                    public boolean onFail(Exception e) {
+//                        Log.e("===", "onFail" + e.getMessage());
+//
+//                        return false;
+//                    }
+//
+//                    @Override
+//                    public void preLoad() {
+//                        Log.e("===", "preLoad" );
+//
+//                    }
+//                }).into(findViewById(R.id.img));
 
-                    @Override
-                    public boolean onFail(Exception e) {
-                        Log.e("===", "onFail" + e.getMessage());
 
-                        return false;
-                    }
+        ImageLoader.with(this).asBitmap().load(url).into(new BitmapTarget() {
+            @Override
+            public void onResourceReady(@NonNull Bitmap resource, @Nullable Transition<? super Bitmap> transition) {
+                Log.e("TAG", "onResourceReady" + resource);
 
-                    @Override
-                    public void preLoad() {
-                        Log.e("===", "preLoad" );
+                ImageView viewById = findViewById(R.id.img);
+                viewById.setImageBitmap(resource);
+            }
 
-                    }
-                }).into(findViewById(R.id.img3));
+            @Override
+            public void onLoadCleared(@Nullable Drawable placeholder) {
 
-        ImageLoader.with(this)
-                .load(url)
-                .rectRoundCorner(1)
-                .openBlur()
-                .asCircle()
-                .placeHolder(R.mipmap.common_course_dufault)
-                .listener(new LoadListener() {
-                    @Override
-                    public boolean onSuccess(Object resource) {
+            }
 
-                        Log.e("===>>>", "onSuccess" + resource);
-                        return false;
-                    }
+            @Override
+            public void onLoadStarted(@Nullable Drawable placeholder) {
+                super.onLoadStarted(placeholder);
+                Log.e("TAG", "onResourceReady start" + placeholder);
 
-                    @Override
-                    public boolean onFail(Exception e) {
-                        Log.e("===", "onFail" + e.getMessage());
 
-                        return false;
-                    }
+            }
 
-                    @Override
-                    public void preLoad() {
-                        Log.e("===", "preLoad" );
+            @Override
+            public void onLoadFailed(@Nullable Drawable errorDrawable) {
+                super.onLoadFailed(errorDrawable);
+                Log.e("TAG", "onResourceReady fail" + errorDrawable);
 
-                    }
-                }).into(findViewById(R.id.img));
-
+            }
+        });
     }
 
     @Override