+++ lib_imageloader/src/main/java/com/nj/baijiayun/imageloader/target/GifTarget.java
@@ -1,5 +1,8 @@
 package com.nj.baijiayun.imageloader.target;
 
+import com.bumptech.glide.load.resource.gif.GifDrawable;
+import com.bumptech.glide.request.target.CustomTarget;
+
 /**
  * @author chengang
  * @date 2019/5/13
@@ -8,5 +11,6 @@ package com.nj.baijiayun.imageloader.target;
  * @package_name com.nj.baijiayun.imageloader.target
  * @describe
  */
-public class GifTarget {
+public abstract  class GifTarget extends CustomTarget<GifDrawable> {
+
 }